pyproject.toml
[project]
name = "py-quickbooks"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohamed Douare"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-quickbooks (>=0.9.12,<0.10.0)",
    "intuit-oauth (>=1.2.6,<2.0.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
